fix: pin terraform version everywhere
Do not blindly allow any version of Terraform. Primarily, we want to
avoid a future confusion when a new incompatible version of Terraform
throws a random weird error. Prefer a well-known error about version
mismatch. Fail fast and clean.

Allow versions 0.12.29 and 0.13.6, because they work with the examples.
We will bump to 0.14 one example at a time.