feat(vm-series): add enable_plan
Not all images work with a plan{} clause, so make it conditional. It
seems necessary for bundle1/2. For some users it may prevent usage of
a custom image, if they are unable to officially accept what seemingly
is just a dummy "purchase".

For all users it prevents deploying a linux for temporary troubleshooting.