feat(vm-series): add Application Insights metrics
Verified that the defaults here do not currently cost anything and do
not require extra permissions. Users are billed only when they actually
manually paste the metrics_instrumentation_key by GB of metrics sent.

An opinionated setup is thus to furnish a single Application Insights
for every module instantiation.

Seemingly, no RBAC changes are required for this to work, it doesn't
matter if firewall runs on a SystemAssigned or a UserAssigned or even
No Identity.