refactor(examples): transit_vnet_common rename
Following the rename of module from "vm-series" to "vmseries" we rename
an example. Since it is quite a complex example for newcomers (they need
to understand for_each, bootstrapping, LB as a next-hop, optionality of
RG and AvSet), I am renaming it to transit_vnet_common. This adheres to
TERRAM-4 story.

The word `common` refers to the Reference Architecture
name being officially "Transit VNET Model (Common Firewall Option)".

The underscores (transit_vnet_common instead of transit-vnet-common)
are for consistency with AWS and GCP repos and with TERRAM-4 story.