docs(examples/vmseries): create basic stub
To avoid confusion, tell everyone where to find usage for module
`vmseries`.

The intention is to first stabilize "transit_vnet_common" and then to
base "vmseries" example on the stripped and minimized version of it.

But I'm removing the previous "standalone" designation. It is the most
basic (and the most used) example, and it will benefit from the shortest
and simplest name. The "vmseries" says it all.