feat(vmseries)!: high_availability always on
Do not allow deployments which do not specify either Availability Zone
or Set. Even if it is one box, we assume user will grow it to two boxes
sooner or later.

The default is Zone "1", which is not usable in some "weak" Regions.

Joining the avzone and avset_id into a single `high_availability` object
because Terraform-0.13.6 cannot validate across two inputs:

Error: Invalid reference in variable validation

  on ../../modules/vmseries/variables.tf line 22, in variable "avzone":
  22:     condition     = var.avzone != null || var.avset_id != null

The condition for variable "avzone" can only refer to the variable itself,
using var.avzone.