fix(examples/tvc): single backend_name
Users do not have anything interesting to tweak in the outbound load
balancer. Removing it from tfvars to greatly simplify code.

The inbound frontends are very interesting to users, to the point that
they may be even confused thinking that "private_frontends" also refer
to the inbound traffic, but to the private IP addressing. (Which is not
the case.) Hence naming the input just "frontend_ips".

Auto-generate resource_group_name.