feat(examples/vmseries): new example, minimal
Create a very minimal example. It's the only one which is 0.12.29
compatible, because it has no module depends_on or for_each anywhere.

As this example is not intended to be copy-pasted into useful code
bases, majority of usual variables are inlined into the main.tf code.
This way a user has everything in place and it aids their learning.