fix(panorama): avoid string-to-bool comparison (#94)
* fix(examples/panorama): input for priv ip

Subnet cidr is an input, but can only be modified if priv ip is not
hardcoded. Making it an input as well.

* feat(examples/panorama): username output

It is more usable if the username/password are printed together after
`terraform apply` completes.

* fix(panorama): remove try()

The `try(1)` is exactly the same as `1`. For the try() to do anything,
it needs to have two or more arguments: `try(1, null)` or `try(1, 2, 3,
4)`

* fix(panorama): avoid string values for bool

Avoid `"false"` which has type `string`, to be compared to `false`, which has
type `bool`. This way if user uses `"False"` capitalized the behavior
will be saner.