feat(examples): Add a minimal `vmseries` example
Create a very minimal example. It's the only one which is 0.12.29
compatible, because it does not employ module `depends_on` or module
`for_each` anywhere.

As this example is not intended to be copy-pasted into maintainable
codebases, a majority of usual variables are inlined into the main.tf
code. This way users see everything together, which can aid their
learning.