feat: add input `enable_zones` to most modules
For provider >= 2.63 it became impossible to create a PublicIP in some
regions, like East Asia. To make them work, I'm adding a new input
`enable_zones`. The current default means nothing changes on the usual
regions like US East 2.

There is no way to also make possible older providers to support
the same code, so I'm bumping the `version` contraints.

fix(panorama):

Before, omitting `zone` on `logging_disks` always resulted in
`Error: expected "zones.0" to not be an empty string` and thus it
had effectively no working default value.

Now the same input will default to the value of input `avzone`,
so that the disk is placed together with the owning virtual machine
and not suspectible to accidental zone separations.