feat(bootstrap): add the new input `files_md5` to facilitate dynamic files
Add the new input `files_md5`. It is optional, to retain the backward
compatibility.

The old problem was that the `files` could not be auto-generated on the
same Apply as this module. The builtin `md5(file(x))`
fails hard if the file does not exist during Validate phase, which is
earlier than Plan or Apply.

For users it is a hurdle on the way to two features:

- to read the secrets out of some vault within Terraform, which would be
very useful for authcodes.
- to read the vm-auth-key from a Panorama.

Now, users are able to generate the `files` on the spot, provided that
they supply the `files_md5` for them. If the `files` entries correspond
to the files that actually exist before the Terraform run, nothing
changes for these (that's the old way used everywhere in the examples).